+++ legadilo/core/forms/widgets.py
@@ -14,11 +14,29 @@ from django.utils.functional import Promise
 class SelectMultipleAutocompleteWidget(widgets.SelectMultiple):
     template_name = "core/widgets/select_multiple_autocomplete.html"
 
-    def __init__(self, attrs=None, choices=(), *, allow_new: bool = True, empty_label=""):
+    def __init__(
+        self,
+        attrs=None,
+        choices=(),
+        *,
+        allow_new: bool = True,
+        empty_label="",
+        server_url: Promise | None = None,
+        extra_server_params: dict[str, Any] | None = None,
+    ):
         attrs = attrs or {}
         attrs["data-bs5-tags"] = "true"
+
         if allow_new:
             attrs["data-allow-new"] = "true"
+
+        if server_url is not None:
+            attrs["data-server"] = server_url
+            attrs["data-live-server"] = "1"
+
+        if extra_server_params:
+            attrs["data-server-params"] = json.dumps(extra_server_params)
+
         super().__init__(attrs, choices)
         self._empty_label = empty_label